RFC. NIST's guidelines on managing TLS. RFC8446 - TLS 1.3 MTLS Bound Access Tokens.
Code.
Replace axum by warp and simplify HTTP request processing.
All we need, ofterntimes, is a simple methods to access and validate
request parameters in the query string. Warp seems to be the right
abtraction that we need at the moment.

Redesigned the routes and redefined handlers. We maintain GET and POST
handlers for 'authorize' request.
    This is just the beginning. We need to modularize the oauth 2.0
    handlers. What we have now is a simple experiment with warp's
    routes, filters, and such.

Removed the dependency on log4j, and started using env_logger,
This is adequate for our purposes.